requirements.txt
flask
line-bot-sdk
openai
python-dotenv
redis
requests
Pillow
